successful conversion of device table

+++ schema/convert/create/USERSLA.sql
@@ -1,6 +1,6 @@
 CREATE TABLE USERSLA (
-  userid int(11) default NULL,
-  slaid int(11) default NULL,
-  start bigint(20) default NULL,
-  end bigint(20) default NULL
+  userid integer default NULL,
+  slaid integer default NULL,
+  start integer default NULL,
+  end bigint default NULL
 );
